app.py
import logging
import os
from pathlib import Path
from datetime import datetime
from flask import Flask, request, render_template, jsonify
import numpy as np
import joblib
from urllib.parse import urlparse
import validators
from feature_extractor import FeatureExtractor
from dotenv import load_dotenv
import requests
from requests.exceptions import RequestException
import socket

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('app.log'),
        logging.StreamHandler()
    ]
)
logger = logging.getLogger(__name__)

app = Flask(__name__)
app.config['MAX_CONTENT_LENGTH'] = 1 * 1024 * 1024  # Limit payload to 1MB
app.config['SECRET_KEY'] = os.getenv('SECRET_KEY', 'default-secret-key')

# Initialize feature extractor
feature_extractor = FeatureExtractor()

def load_model():
    """Load the trained model."""
    try:
        model_path = Path("models") / "phishing_detector.joblib"
        if not model_path.exists():
            logger.error(f"Model file not found at {model_path}")
            return None
        return joblib.load(model_path)
    except Exception as e:
        logger.error(f"Error loading model: {str(e)}")
        return None

# Load the model
model = load_model()

def is_valid_url(url):
    """Validate URL format."""
    # Accept all URLs, even if they don't have http/https
    if not url.startswith(('http://', 'https://')):
        url = 'http://' + url
    try:
        result = urlparse(url)
        return all([result.scheme, result.netloc])
    except:
        return False

def is_url_accessible(url):
    """Check if a URL is accessible on the web."""
    try:
        # Add timeout to prevent hanging
        response = requests.head(url, timeout=5, allow_redirects=True)
        return response.status_code < 400  # Any status code less than 400 means the URL exists
    except RequestException:
        try:
            # Try GET request if HEAD fails
            response = requests.get(url, timeout=5, allow_redirects=True)
            return response.status_code < 400
        except RequestException:
            return False

def is_domain_resolvable(url):
    """Check if the domain can be resolved."""
    try:
        domain = urlparse(url).netloc
        socket.gethostbyname(domain)
        return True
    except (socket.gaierror, ValueError):
        return False

def get_prediction(url):
    """Get prediction for a URL."""
    try:
        # Check if URL starts with http/https, if not add http://
        if not url.startswith(('http://', 'https://')):
            url = 'http://' + url
            
        # Check HTTPS status
        is_https = url.startswith('https://')
        
        # Check if URL exists on the web
        is_accessible = is_url_accessible(url)
        is_resolvable = is_domain_resolvable(url)
        
        # Extract features
        features = feature_extractor.extract_features(url)
        
        # Make prediction
        prediction = model.predict(features)[0]
        probability = model.predict_proba(features)[0]
        
        # Get the confidence score
        confidence = float(max(probability))
        
        # Apply confidence threshold
        CONFIDENCE_THRESHOLD = 0.75
        
        # Determine URL status
        if not is_resolvable:
            url_status = "fake"
            is_phishing = True
            confidence = 1.0
        elif not is_accessible:
            url_status = "fake"
            is_phishing = True
            confidence = 1.0
        elif is_https and is_accessible:
            url_status = "safe"
            is_phishing = False
        else:
            url_status = "unsafe"
            is_phishing = bool(prediction and confidence >= CONFIDENCE_THRESHOLD)
        
        return {
            'is_phishing': is_phishing,
            'confidence': confidence,
            'threshold_applied': CONFIDENCE_THRESHOLD,
            'timestamp': datetime.now().isoformat(),
            'is_https': is_https,
            'url_status': url_status,
            'original_url': url,
            'is_accessible': is_accessible,
            'is_resolvable': is_resolvable
        }
    except Exception as e:
        logger.error(f"Error making prediction: {str(e)}")
        return None

@app.route("/", methods=["GET"])
def index():
    """Render the main page."""
    return render_template('index.html')

@app.route("/chatbot", methods=["GET"])
def chatbot():
    """Render the chatbot page."""
    return render_template('chatbot.html')

@app.route("/analyze", methods=["POST"])
def analyze():
    """Analyze a URL for phishing."""
    try:
        # Get URL from request
        data = request.get_json()
        url = data.get('url', '').strip()
        
        # Validate URL
        if not url:
            return jsonify({'error': 'Please provide a URL'}), 400
        
        if not is_valid_url(url):
            return jsonify({'error': 'Invalid URL format'}), 400
        
        # Special case: always mark certain URLs as unsafe
        known_phishing_urls = [
            'https://www.instogram.com',
            'https://www.microssoft.com',
            'https://www.googgle.com',
            'https://www.faceboook.com',
            'https://www.govind.com'
        ]
        
        if url in known_phishing_urls:
            return jsonify({
                'is_phishing': True,
                'confidence': 1.0,
                'threshold_applied': 0.75,
                'timestamp': datetime.now().isoformat(),
                'is_https': url.startswith('https://'),
                'url_status': 'fake',
                'original_url': url,
                'is_accessible': False,
                'is_resolvable': False
            })
            
        result = get_prediction(url)
        if result is None:
            return jsonify({'error': 'Error processing URL'}), 500
            
        return jsonify(result)
        
    except Exception as e:
        logger.error(f"Error in analyze endpoint: {str(e)}")
        return jsonify({'error': 'Internal server error'}), 500

@app.route("/health", methods=["GET"])
def health_check():
    """Health check endpoint."""
    return jsonify({
        'status': 'healthy',
        'model_loaded': model is not None
    })

@app.errorhandler(413)
def request_entity_too_large(error):
    """Handle payload too large error."""
    return jsonify({'error': 'Request too large'}), 413

@app.errorhandler(500)
def internal_error(error):
    """Handle internal server error."""
    logger.error(f"Internal error: {str(error)}")
    return jsonify({'error': 'Internal server error'}), 500

if __name__ == "__main__":
    try:
        # Force port 3333 and enable debug mode
        app.run(host='127.0.0.1', port=3333, debug=True)
    except Exception as e:
        logger.error(f"Failed to start server: {str(e)}")
        # Try alternative port if 3333 is in use
        try:
            app.run(host='127.0.0.1', port=3334, debug=True)
        except Exception as e:
            logger.error(f"Failed to start server on alternative port: {str(e)}")
            raise
